rename and move store copy process class
It was badly named, placed and instantiated.
